chore: project structure reorganization, renaming
