+++ src/app/app.routes.ts
@@ -1,11 +1,11 @@
 import { Routes } from '@angular/router';
-import { HomepageComponent } from '../components/homepage/homepage.component';
-import { AboutMePageComponent } from '../components/about-me-page/about-me-page.component';
-import { TimelinePageComponent } from '../components/timeline-page/timeline-page.component';
+import { HomepageComponent } from '../pages/homepage/homepage.component';
+import { AboutMeComponent } from '../pages/about-me/about-me.component';
+import { TimelineComponent } from '../pages/timeline/timeline.component';
 
 export const routes: Routes = [
   { path: '', component: HomepageComponent },
-  { path: 'about-me', component: AboutMePageComponent },
-  { path: 'timeline', component: TimelinePageComponent },
+  { path: 'about-me', component: AboutMeComponent },
+  { path: 'timeline', component: TimelineComponent },
   { path: '**', redirectTo: '' },
 ];
