+++ src/components/header/header.component.html
@@ -1,7 +1,8 @@
 <div class="header">
   <!--Name container -->
   <div class="name-container">
-    <a class="name-text" routerLink="">Pritam Kantane</a>
+    <a class="name-text" routerLink="">PyyLog</a>
+    <p class="name-subtitle">(Pritam Kantane)</p>
   </div>
 
   <!--Social networks container -->
