+++ src/components/header/header.component.scss
@@ -1,4 +1,4 @@
-@import url('https://fonts.googleapis.com/css2?family=Rubik:ital,wght@0,300..900;1,300..900&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Pattaya&display=swap');
 
 .header {
   display: flex;
@@ -15,17 +15,32 @@
 .name-container {
   display: flex;
   position: absolute;
+  flex-direction: column;
   justify-content: center;
   align-items: center;
-  height: 6rem;
   width: 100%;
 
   .name-text {
     color: white;
-    font-size: 1.5rem;
-    font-family: 'Permanent Marker', cursive;
+    font-size: 2rem;
+    letter-spacing: 0.1rem;
+    font-family: 'Pattaya', sans-serif;
     text-decoration: none;
     cursor: pointer;
+  }
+
+  .name-subtitle {
+    position: absolute;
+    color: white;
+    font-size: 1rem;
+    font-family: 'Rubik', sans-serif;
+    opacity: 0;
+    top: 1rem;
+  }
+
+  &:hover .name-subtitle {
+    opacity: 1;
+    top: 1.75rem;
   }
 }
