+++ src/pages/homepage/homepage.component.html
@@ -77,7 +77,7 @@
         </div>
 
         <div class="path-to-code">
-          <a *ngFor="let path of selectedProject.pathToCodes; let i = index" href="{{ path }}">
+          <a *ngFor="let path of selectedProject.pathToCodes; let i = index" href="{{ path }}" target="_blank">
             <span>{{ selectedProject.spanNames[i] }}</span>
           </a>
         </div>
@@ -96,9 +96,9 @@
           <div class="contact-text-container">
             <p class="contact-text">
               No form here 😎, just a magic link to my
-              <a class="linkedin-contact-text" [href]="links.linkedin">LinkedIn</a>. When sending a DM, add the following sentence:
-              <span class="implicit-indicator">"That shooting star was too fast !"</span>
-              so that I know you came from here.🙂
+              <a class="linkedin-contact-text" [href]="links.linkedin" target="_blank">LinkedIn</a>. When sending a DM, add the following 'code':
+              <span class="implicit-indicator">"ICFYP"</span>
+              so that I know you came from here (or simply mention it) ! 🙂
             </p>
           </div>
         </li>
