+++ src/pages/homepage/homepage.component.scss
@@ -362,7 +362,7 @@
     width: 100%;
     list-style-type: none;
     border-bottom: 0.1rem solid white;
-    padding: 1rem 1rem;
+    padding: 1rem 2rem;
     text-align: center;
   }
